optimization in filter selector

+++ modules/siddhi-core/src/main/java/org/wso2/siddhi/core/gpu/query/selector/GpuQuerySelector.java
@@ -148,7 +148,7 @@ public class GpuQuerySelector extends QuerySelector {
                 for (GpuEventAttribute attrib : gpuMetaEventAttributeList) {
                     switch(attrib.type) {
                     case BOOL:
-                        attributeData[index++] = outputEventBuffer.getShort();
+                        attributeData[index++] = Short.valueOf(outputEventBuffer.getShort());
                         break;
                     case INT:
                         attributeData[index++] = outputEventBuffer.getInt();
